+++ src/main/java/org/sagebionetworks/web/client/widget/entity/tabs/Tabs.java
@@ -64,8 +64,4 @@ public class Tabs implements IsWidget {
   public void setNavTabsVisible(boolean visible) {
     view.setNavTabsVisible(visible);
   }
-
-  public void setEntitySidebar(String entityId, Double versionNumber) {
-    view.setEntitySidebar(entityId, versionNumber);
-  }
 }
