+++ src/main/java/org/sagebionetworks/web/client/widget/entity/tabs/TabsView.java
@@ -8,6 +8,4 @@ public interface TabsView extends IsWidget {
   void clear();
 
   void setNavTabsVisible(boolean visible);
-
-  void setEntitySidebar(String entityId, Double versionNumber);
 }
