+++ src/main/resources/org/sagebionetworks/web/client/widget/entity/tabs/TabsViewImpl.ui.xml
@@ -4,8 +4,6 @@
   xmlns:ui="urn:ui:com.google.gwt.uibinder"
   xmlns:b="urn:import:org.gwtbootstrap3.client.ui"
   xmlns:bh="urn:import:org.gwtbootstrap3.client.ui.html"
-  xmlns:w="urn:import:org.sagebionetworks.web.client.widget"
-
 >
   <bh:Div>
     <!-- Tabs -->
@@ -17,12 +15,9 @@
       <!-- list of b:TabListItem-->
       <!-- <b:TabListItem ui:field="tabItem1" active="true" dataTarget="#tab1" text="Tab1"/>  -->
     </b:NavTabs>
-    <bh:Div addStyleNames="entityPageContentRow">
-      <b:TabContent ui:field="tabContent">
-        <!-- list of b:TabPane-->
-        <!-- <b:TabPane active="true" addStyleNames="margin-top-15"> -->
-      </b:TabContent>
-       <w:ReactComponent ui:field="entitySidebar" addStyleNames="entityPageSidebar"/>
-    </bh:Div>
+    <b:TabContent ui:field="tabContent">
+      <!-- list of b:TabPane-->
+      <!-- <b:TabPane active="true" addStyleNames="margin-top-15"> -->
+    </b:TabContent>
   </bh:Div>
 </ui:UiBinder>
